Adding setting to skip of automatic pgviews sync
This setting allows a dev to disable the automatic syncing behavior.
I've encountered a situation in a project's continuous integration where the post migration signal is run before some of
the tables the view refers to have been created/migrated. This only happens on a brand new environment and most
glaringly on test servers that might not even run all the migrations for unit tests.